+++ src/Http/Controllers/ParagraphController.php
@@ -10,6 +10,7 @@ use Illuminate\Http\Request;
 use Illuminate\Contracts\View\View;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
+use Velor\Pages\Resources\ParagraphResource;
 use Velor\Pages\Http\Requests\ParagraphRequest;
 use App\Services\Resources\Contracts\ResourceIndexQueryInterface;
 
@@ -17,6 +18,7 @@ class ParagraphController extends Controller
 {
     public function __construct(
         protected ResourceIndexQueryInterface $resourceIndexQuery,
+        protected ParagraphResource $paragraphResource,
     ) {
     }
 
@@ -26,14 +28,12 @@ class ParagraphController extends Controller
             'cms.layouts.index',
             [
                 'pagination' => $this->resourceIndexQuery->paginate(
-                    model: Paragraph::class,
+                    resource: $this->paragraphResource,
                     parent: $page,
                     relationship: 'paragraphs',
                     search: $request->string('search')->toString(),
                 ),
-                'model' => new Paragraph([
-                    'page_id' => $page->getKey(),
-                ]),
+                'resource' => $this->paragraphResource,
             ]
         );
     }
@@ -41,9 +41,7 @@ class ParagraphController extends Controller
     public function create(Page $page): View
     {
         return view('cms.layouts.form', [
-           'model' => new Paragraph([
-               'page_id' => $page->getKey(),
-           ]),
+            'resource' => $this->paragraphResource,
         ]);
     }
 
@@ -59,14 +57,14 @@ class ParagraphController extends Controller
     public function show(Page $page, Paragraph $paragraph): View
     {
         return view('cms.layouts.show', [
-            'model' => $paragraph,
+            'resource' => $this->paragraphResource,
         ]);
     }
 
     public function edit(Page $page, Paragraph $paragraph): View
     {
         return view('cms.layouts.form', [
-            'model' => $paragraph,
+            'resource' => $this->paragraphResource,
         ]);
     }